Split DevicesController: extract state-callbacks mixin
Move the six per-attribute monitor callbacks (_on_state_change,
_on_ip_change, _on_version_change, _on_mac_address_change,
_on_api_encryption_change, _on_config_hash_change) into a
sibling DeviceStateCallbacksMixin in state_callbacks.py.

Same mixin shape as metadata.py (#702): the methods only touch
controller-internal helpers and self._db; no external state
threading needed. Inheritance keeps test instance-patches and
the bound-method captures in DeviceStateMonitor's __init__
working unchanged.

No test changes needed; tests call the methods as bound methods
which still resolve via inheritance.

Pure refactor; behaviour preserved. Drops ~184 lines from
controllers/devices/controller.py (now 1827, finally under
2000).
